+++ AreaProg.AspNetCore.Migrations.Demo/Migrations/V1_0_0_InitialSetup.cs
@@ -1,6 +1,6 @@
 using AreaProg.AspNetCore.Migrations.Demo.Data;
 using AreaProg.AspNetCore.Migrations.Demo.Data.Entities;
-using AreaProg.AspNetCore.Migrations.Models;
+using AreaProg.AspNetCore.Migrations.Abstractions;
 using Microsoft.Extensions.Logging;
 
 namespace AreaProg.AspNetCore.Migrations.Demo.Migrations;